+++ pal.py
@@ -22,8 +22,7 @@ class PAL(dqn.DQN):
         qout = self.q_function(batch_state, test=False)
 
         batch_actions = Variable(
-            self.xp.asarray(
-                [elem['action'] for elem in experiences], dtype=np.int32))
+            self.xp.asarray([elem['action'] for elem in experiences]))
         batch_q = qout.evaluate_actions(batch_actions)
 
         # Compute target values
